+++ tests/test_viz.py
@@ -6,6 +6,7 @@ from pyriemann.utils.viz import (
     plot_confusion_matrix,
     plot_embedding,
     plot_cospectra,
+    plot_waveforms
 )
 
 
@@ -36,3 +37,16 @@ def test_cospectra():
     cosp = np.random.randn(n_freqs, n_channels, n_channels)
     freqs = np.random.randn(n_freqs)
     plot_cospectra(cosp, freqs)
+
+
+@requires_matplotlib
+@pytest.mark.parametrize("display", ["all", "mean", "mean+/-std", "hist"])
+def test_plot_waveforms(display):
+    """Test plot_waveforms"""
+    n_matrices, n_channels, n_times = 16, 3, 50
+    X = np.random.randn(n_matrices, n_channels, n_times)
+    plot_waveforms(X, display)
+    plot_waveforms(X, display, times=np.arange(n_times))
+
+    X = np.random.randn(n_matrices, 1, n_times)
+    plot_waveforms(X, display)